refactor: rename `health` to `probes`
