+++ agent_api_rest/postman/ssi-agent.postman_collection.json
@@ -1521,6 +1521,28 @@
 					"response": []
 				}
 			]
+		},
+		{
+			"name": "_monitoring",
+			"item": [
+				{
+					"name": "Liveness probe",
+					"request": {
+						"method": "GET",
+						"header": [],
+						"url": {
+							"raw": "{{HOST}}/healthz",
+							"host": [
+								"{{HOST}}"
+							],
+							"path": [
+								"healthz"
+							]
+						}
+					},
+					"response": []
+				}
+			]
 		}
 	],
 	"event": [
